+++ packages/model/src/resolver/DevUi.ts
@@ -1,0 +1,3 @@
+import { ElPlusModel } from "./ElementPlus";
+
+export let DevUiModel = ElPlusModel;