+++ packages/model/src/resolver/Vant.ts
@@ -1,0 +1,2 @@
+import { NutUiModel } from "./NutUi";
+export let VantModel = NutUiModel;