+++ packages/model/src/resolver/index.ts
@@ -1,1 +1,6 @@
 export * from "./ElementPlus";
+export * from "./NaiveUi";
+export * from "./NutUi";
+export * from "./AntDesign";
+export * from "./Vant";
+export * from "./DevUi";