+++ packages/model/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vformore/model
 
+## 1.0.6
+
+### Patch Changes
+
+- feature:add more component resolver
+
 ## 1.0.5
 
 ### Patch Changes